+++ python/pyspark/sql/dataframe.py
@@ -15,13 +15,8 @@
 # limitations under the License.
 #
 
-import json
-import os
-import sys
-import random
-import warnings
-from collections.abc import Iterable
-from functools import reduce
+# mypy: disable-error-code="empty-body"
+
 from typing import (
     Any,
     Callable,
@@ -31,45 +26,26 @@ from typing import (
     Optional,
     Sequence,
     Tuple,
-    Type,
     Union,
-    cast,
     overload,
     TYPE_CHECKING,
 )
 
 from pyspark import _NoValue
 from pyspark._globals import _NoValueType
-from pyspark.errors import (
-    PySparkTypeError,
-    PySparkValueError,
-    PySparkIndexError,
-    PySparkAttributeError,
-)
-from pyspark.util import (
-    is_remote_only,
-    _load_from_socket,
-    _local_iterator_from_socket,
-)
-from pyspark.serializers import BatchedSerializer, CPickleSerializer, UTF8Deserializer
+from pyspark.util import is_remote_only
 from pyspark.storagelevel import StorageLevel
-from pyspark.traceback_utils import SCCallSiteSync
-from pyspark.sql.column import Column, _to_seq, _to_list, _to_java_column
+from pyspark.resource import ResourceProfile
+from pyspark.sql.column import Column
 from pyspark.sql.readwriter import DataFrameWriter, DataFrameWriterV2
 from pyspark.sql.streaming import DataStreamWriter
-from pyspark.sql.types import (
-    StructType,
-    Row,
-    _parse_datatype_json_string,
-)
-from pyspark.sql.utils import get_active_spark_context, toJArray
-from pyspark.sql.pandas.conversion import PandasConversionMixin
-from pyspark.sql.pandas.map_ops import PandasMapOpsMixin
+from pyspark.sql.types import StructType, Row
+from pyspark.sql.utils import dispatch_df_method
 
 if TYPE_CHECKING:
     from py4j.java_gateway import JavaObject
-    from pyspark.core.rdd import RDD
     from pyspark.core.context import SparkContext
+    from pyspark.core.rdd import RDD
     from pyspark._typing import PrimitiveType
     from pyspark.pandas.frame import DataFrame as PandasOnSparkDataFrame
     from pyspark.sql._typing import (
@@ -82,12 +58,17 @@ if TYPE_CHECKING:
     from pyspark.sql.session import SparkSession
     from pyspark.sql.group import GroupedData
     from pyspark.sql.observation import Observation
+    from pyspark.sql.pandas._typing import (
+        PandasMapIterFunction,
+        ArrowMapIterFunction,
+        DataFrameLike as PandasDataFrameLike,
+    )
 
 
 __all__ = ["DataFrame", "DataFrameNaFunctions", "DataFrameStatFunctions"]
 
 
-class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
+class DataFrame:
     """A distributed collection of data grouped into named columns.
 
     .. versionadded:: 1.3.0
@@ -125,12 +106,13 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
 
     >>> people.filter(people.age > 30).join(
     ...     department, people.deptId == department.id).groupBy(
-    ...     department.name, "gender").agg({"salary": "avg", "age": "max"}).show()
+    ...     department.name, "gender").agg(
+    ...         {"salary": "avg", "age": "max"}).sort("max(age)").show()
     +-------+------+-----------+--------+
     |   name|gender|avg(salary)|max(age)|
     +-------+------+-----------+--------+
-    |     ML|     F|      150.0|      60|
     |PySpark|     M|       75.0|      50|
+    |     ML|     F|      150.0|      60|
     +-------+------+-----------+--------+
 
     Notes
@@ -139,49 +121,26 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     created via using the constructor.
     """
 
-    def __init__(
-        self,
+    # HACK ALERT!! this is to reduce the backward compatibility concern, and returns
+    # Spark Classic DataFrame by default. This is NOT an API, and NOT supposed to
+    # be directly invoked. DO NOT use this constructor.
+    _sql_ctx: Optional["SQLContext"]
+    _session: "SparkSession"
+    _sc: "SparkContext"
+    _jdf: "JavaObject"
+    is_cached: bool
+    _schema: Optional[StructType]
+    _lazy_rdd: Optional["RDD[Row]"]
+    _support_repr_html: bool
+
+    def __new__(
+        cls,
         jdf: "JavaObject",
         sql_ctx: Union["SQLContext", "SparkSession"],
-    ):
-        from pyspark.sql.context import SQLContext
+    ) -> "DataFrame":
+        from pyspark.sql.classic.dataframe import DataFrame
 
-        self._sql_ctx: Optional["SQLContext"] = None
-
-        if isinstance(sql_ctx, SQLContext):
-            assert not os.environ.get("SPARK_TESTING")  # Sanity check for our internal usage.
-            assert isinstance(sql_ctx, SQLContext)
-            # We should remove this if-else branch in the future release, and rename
-            # sql_ctx to session in the constructor. This is an internal code path but
-            # was kept with a warning because it's used intensively by third-party libraries.
-            warnings.warn("DataFrame constructor is internal. Do not directly use it.")
-            self._sql_ctx = sql_ctx
-            session = sql_ctx.sparkSession
-        else:
-            session = sql_ctx
-        self._session: "SparkSession" = session
-
-        self._sc: "SparkContext" = sql_ctx._sc
-        self._jdf: "JavaObject" = jdf
-        self.is_cached = False
-        # initialized lazily
-        self._schema: Optional[StructType] = None
-        self._lazy_rdd: Optional["RDD[Row]"] = None
-        # Check whether _repr_html is supported or not, we use it to avoid calling _jdf twice
-        # by __repr__ and _repr_html_ while eager evaluation opens.
-        self._support_repr_html = False
-
-    @property
-    def sql_ctx(self) -> "SQLContext":
-        from pyspark.sql.context import SQLContext
-
-        warnings.warn(
-            "DataFrame.sql_ctx is an internal property, and will be removed "
-            "in future releases. Use DataFrame.sparkSession instead."
-        )
-        if self._sql_ctx is None:
-            self._sql_ctx = SQLContext._get_or_create(self._sc)
-        return self._sql_ctx
+        return DataFrame.__new__(DataFrame, jdf, sql_ctx)
 
     @property
     def sparkSession(self) -> "SparkSession":
@@ -202,7 +161,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> type(df.sparkSession)
         <class '...session.SparkSession'>
         """
-        return self._session
+        ...
 
     if not is_remote_only():
 
@@ -222,14 +181,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
             >>> type(df.rdd)
             <class 'pyspark.core.rdd.RDD'>
             """
-            from pyspark.core.rdd import RDD
-
-            if self._lazy_rdd is None:
-                jrdd = self._jdf.javaToPython()
-                self._lazy_rdd = RDD(
-                    jrdd, self.sparkSession._sc, BatchedSerializer(CPickleSerializer())
-                )
-            return self._lazy_rdd
+            ...
 
     @property
     def na(self) -> "DataFrameNaFunctions":
@@ -259,7 +211,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  1|  2|
         +---+---+
         """
-        return DataFrameNaFunctions(self)
+        ...
 
     @property
     def stat(self) -> "DataFrameStatFunctions":
@@ -283,10 +235,11 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.stat.corr("id", "c")
         1.0
         """
-        return DataFrameStatFunctions(self)
+        ...
 
     if not is_remote_only():
 
+        @dispatch_df_method
         def toJSON(self, use_unicode: bool = True) -> "RDD[str]":
             """Converts a :class:`DataFrame` into a :class:`RDD` of string.
 
@@ -309,11 +262,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
             >>> df.toJSON().first()
             '{"age":2,"name":"Alice"}'
             """
-            from pyspark.core.rdd import RDD
+            ...
 
-            rdd = self._jdf.toJSON()
-            return RDD(rdd.toJavaRDD(), self._sc, UTF8Deserializer(use_unicode))
-
+    @dispatch_df_method
     def registerTempTable(self, name: str) -> None:
         """Registers this :class:`DataFrame` as a temporary table using the given name.
 
@@ -344,9 +295,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         True
 
         """
-        warnings.warn("Deprecated in 2.0, use createOrReplaceTempView instead.", FutureWarning)
-        self._jdf.createOrReplaceTempView(name)
+        ...
 
+    @dispatch_df_method
     def createTempView(self, name: str) -> None:
         """Creates a local temporary view with this :class:`DataFrame`.
 
@@ -410,8 +361,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  4|Jill|
         +---+----+
         """
-        self._jdf.createTempView(name)
+        ...
 
+    @dispatch_df_method
     def createOrReplaceTempView(self, name: str) -> None:
         """Creates or replaces a local temporary view with this :class:`DataFrame`.
 
@@ -453,8 +405,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ... spark.catalog.dropTempView("people")
         True
         """
-        self._jdf.createOrReplaceTempView(name)
+        ...
 
+    @dispatch_df_method
     def createGlobalTempView(self, name: str) -> None:
         """Creates a global temporary view with this :class:`DataFrame`.
 
@@ -501,8 +454,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> spark.catalog.dropGlobalTempView("people")
         True
         """
-        self._jdf.createGlobalTempView(name)
+        ...
 
+    @dispatch_df_method
     def createOrReplaceGlobalTempView(self, name: str) -> None:
         """Creates or replaces a global temporary view using the given name.
 
@@ -537,7 +491,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> spark.catalog.dropGlobalTempView("people")
         True
         """
-        self._jdf.createOrReplaceGlobalTempView(name)
+        ...
 
     @property
     def write(self) -> DataFrameWriter:
@@ -566,7 +520,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.write.saveAsTable("tab2")
         >>> _ = spark.sql("DROP TABLE tab2")
         """
-        return DataFrameWriter(self)
+        ...
 
     @property
     def writeStream(self) -> DataStreamWriter:
@@ -602,7 +556,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...     time.sleep(3)
         ...     query.stop()
         """
-        return DataStreamWriter(self)
+        ...
 
     @property
     def schema(self) -> StructType:
@@ -646,18 +600,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         StructType([StructField('value', StringType(), False)])
 
         """
-        if self._schema is None:
-            try:
-                self._schema = cast(
-                    StructType, _parse_datatype_json_string(self._jdf.schema().json())
-                )
-            except Exception as e:
-                raise PySparkValueError(
-                    error_class="CANNOT_PARSE_DATATYPE",
-                    message_parameters={"error": str(e)},
-                )
-        return self._schema
+        ...
 
+    @dispatch_df_method
     def printSchema(self, level: Optional[int] = None) -> None:
         """Prints out the schema in the tree format.
         Optionally allows to specify how many levels to print if schema is nested.
@@ -710,11 +655,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
          |-- nonnullable: long (nullable = false)
          |-- nullable: void (nullable = true)
         """
-        if level:
-            print(self._jdf.schema().treeString(level))
-        else:
-            print(self._jdf.schema().treeString())
+        ...
 
+    @dispatch_df_method
     def explain(
         self, extended: Optional[Union[bool, str]] = None, mode: Optional[str] = None
     ) -> None:
@@ -782,56 +725,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...Statistics...
         ...
         """
+        ...
 
-        if extended is not None and mode is not None:
-            raise PySparkValueError(
-                error_class="CANNOT_SET_TOGETHER",
-                message_parameters={"arg_list": "extended and mode"},
-            )
-
-        # For the no argument case: df.explain()
-        is_no_argument = extended is None and mode is None
-
-        # For the cases below:
-        #   explain(True)
-        #   explain(extended=False)
-        is_extended_case = isinstance(extended, bool) and mode is None
-
-        # For the case when extended is mode:
-        #   df.explain("formatted")
-        is_extended_as_mode = isinstance(extended, str) and mode is None
-
-        # For the mode specified:
-        #   df.explain(mode="formatted")
-        is_mode_case = extended is None and isinstance(mode, str)
-
-        if not (is_no_argument or is_extended_case or is_extended_as_mode or is_mode_case):
-            if (extended is not None) and (not isinstance(extended, (bool, str))):
-                raise PySparkTypeError(
-                    error_class="NOT_BOOL_OR_STR",
-                    message_parameters={
-                        "arg_name": "extended",
-                        "arg_type": type(extended).__name__,
-                    },
-                )
-            if (mode is not None) and (not isinstance(mode, str)):
-                raise PySparkTypeError(
-                    error_class="NOT_STR",
-                    message_parameters={"arg_name": "mode", "arg_type": type(mode).__name__},
-                )
-
-        # Sets an explain mode depending on a given argument
-        if is_no_argument:
-            explain_mode = "simple"
-        elif is_extended_case:
-            explain_mode = "extended" if extended else "simple"
-        elif is_mode_case:
-            explain_mode = cast(str, mode)
-        elif is_extended_as_mode:
-            explain_mode = cast(str, extended)
-        assert self._sc._jvm is not None
-        print(self._sc._jvm.PythonSQLUtils.explainString(self._jdf.queryExecution(), explain_mode))
-
+    @dispatch_df_method
     def exceptAll(self, other: "DataFrame") -> "DataFrame":
         """Return a new :class:`DataFrame` containing rows in this :class:`DataFrame` but
         not in another :class:`DataFrame` while preserving duplicates.
@@ -869,8 +765,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         +---+---+
 
         """
-        return DataFrame(self._jdf.exceptAll(other._jdf), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def isLocal(self) -> bool:
         """Returns ``True`` if the :func:`collect` and :func:`take` methods can be run locally
         (without any Spark executors).
@@ -890,7 +787,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.isLocal()
         True
         """
-        return self._jdf.isLocal()
+        ...
 
     @property
     def isStreaming(self) -> bool:
@@ -921,8 +818,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.isStreaming
         True
         """
-        return self._jdf.isStreaming()
+        ...
 
+    @dispatch_df_method
     def isEmpty(self) -> bool:
         """
         Checks if the :class:`DataFrame` is empty and returns a boolean value.
@@ -972,8 +870,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df_no_rows.isEmpty()
         True
         """
-        return self._jdf.isEmpty()
+        ...
 
+    @dispatch_df_method
     def show(self, n: int = 20, truncate: Union[bool, int] = True, vertical: bool = False) -> None:
         """
         Prints the first ``n`` rows of the DataFrame to the console.
@@ -1063,65 +962,21 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         age  | 19
         name | This is a super l...
         """
-        print(self._show_string(n, truncate, vertical))
+        ...
 
-    def _show_string(
-        self, n: int = 20, truncate: Union[bool, int] = True, vertical: bool = False
-    ) -> str:
-        if not isinstance(n, int) or isinstance(n, bool):
-            raise PySparkTypeError(
-                error_class="NOT_INT",
-                message_parameters={"arg_name": "n", "arg_type": type(n).__name__},
-            )
-
-        if not isinstance(vertical, bool):
-            raise PySparkTypeError(
-                error_class="NOT_BOOL",
-                message_parameters={"arg_name": "vertical", "arg_type": type(vertical).__name__},
-            )
-
-        if isinstance(truncate, bool) and truncate:
-            return self._jdf.showString(n, 20, vertical)
-        else:
-            try:
-                int_truncate = int(truncate)
-            except ValueError:
-                raise PySparkTypeError(
-                    error_class="NOT_BOOL",
-                    message_parameters={
-                        "arg_name": "truncate",
-                        "arg_type": type(truncate).__name__,
-                    },
-                )
-
-            return self._jdf.showString(n, int_truncate, vertical)
-
+    @dispatch_df_method
     def __repr__(self) -> str:
-        if not self._support_repr_html and self.sparkSession._jconf.isReplEagerEvalEnabled():
-            vertical = False
-            return self._jdf.showString(
-                self.sparkSession._jconf.replEagerEvalMaxNumRows(),
-                self.sparkSession._jconf.replEagerEvalTruncate(),
-                vertical,
-            )
-        else:
-            return "DataFrame[%s]" % (", ".join("%s: %s" % c for c in self.dtypes))
+        ...
 
+    @dispatch_df_method
     def _repr_html_(self) -> Optional[str]:
         """Returns a :class:`DataFrame` with html code when you enabled eager evaluation
         by 'spark.sql.repl.eagerEval.enabled', this only called by REPL you are
         using support eager evaluation with HTML.
         """
-        if not self._support_repr_html:
-            self._support_repr_html = True
-        if self.sparkSession._jconf.isReplEagerEvalEnabled():
-            return self._jdf.htmlString(
-                self.sparkSession._jconf.replEagerEvalMaxNumRows(),
-                self.sparkSession._jconf.replEagerEvalTruncate(),
-            )
-        else:
-            return None
+        ...
 
+    @dispatch_df_method
     def checkpoint(self, eager: bool = True) -> "DataFrame":
         """Returns a checkpointed version of this :class:`DataFrame`. Checkpointing can be used to
         truncate the logical plan of this :class:`DataFrame`, which is especially useful in
@@ -1154,9 +1009,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...     df.checkpoint(False)
         DataFrame[age: bigint, name: string]
         """
-        jdf = self._jdf.checkpoint(eager)
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def localCheckpoint(self, eager: bool = True) -> "DataFrame":
         """Returns a locally checkpointed version of this :class:`DataFrame`. Checkpointing can be
         used to truncate the logical plan of this :class:`DataFrame`, which is especially useful in
@@ -1186,9 +1041,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.localCheckpoint(False)
         DataFrame[age: bigint, name: string]
         """
-        jdf = self._jdf.localCheckpoint(eager)
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def withWatermark(self, eventTime: str, delayThreshold: str) -> "DataFrame":
         """Defines an event time watermark for this :class:`DataFrame`. A watermark tracks a point
         in time before which we assume no more late data is going to arrive.
@@ -1252,22 +1107,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> time.sleep(3)
         >>> query.stop()
         """
-        if not eventTime or type(eventTime) is not str:
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "eventTime", "arg_type": type(eventTime).__name__},
-            )
-        if not delayThreshold or type(delayThreshold) is not str:
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={
-                    "arg_name": "delayThreshold",
-                    "arg_type": type(delayThreshold).__name__,
-                },
-            )
-        jdf = self._jdf.withWatermark(eventTime, delayThreshold)
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def hint(
         self, name: str, *parameters: Union["PrimitiveType", "Column", List["PrimitiveType"]]
     ) -> "DataFrame":
@@ -1308,67 +1150,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ... +- BroadcastHashJoin ...
         ...
         """
-        if len(parameters) == 1 and isinstance(parameters[0], list):
-            parameters = parameters[0]  # type: ignore[assignment]
+        ...
 
-        if not isinstance(name, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "name", "arg_type": type(name).__name__},
-            )
-
-        allowed_types = (str, float, int, Column, list)
-        allowed_primitive_types = (str, float, int)
-        allowed_types_repr = ", ".join(
-            [t.__name__ for t in allowed_types[:-1]]
-            + ["list[" + t.__name__ + "]" for t in allowed_primitive_types]
-        )
-        for p in parameters:
-            if not isinstance(p, allowed_types):
-                raise PySparkTypeError(
-                    error_class="DISALLOWED_TYPE_FOR_CONTAINER",
-                    message_parameters={
-                        "arg_name": "parameters",
-                        "arg_type": type(parameters).__name__,
-                        "allowed_types": allowed_types_repr,
-                        "item_type": type(p).__name__,
-                    },
-                )
-            if isinstance(p, list):
-                if not all(isinstance(e, allowed_primitive_types) for e in p):
-                    raise PySparkTypeError(
-                        error_class="DISALLOWED_TYPE_FOR_CONTAINER",
-                        message_parameters={
-                            "arg_name": "parameters",
-                            "arg_type": type(parameters).__name__,
-                            "allowed_types": allowed_types_repr,
-                            "item_type": type(p).__name__ + "[" + type(p[0]).__name__ + "]",
-                        },
-                    )
-
-        def _converter(parameter: Union[str, list, float, int, Column]) -> Any:
-            if isinstance(parameter, Column):
-                return _to_java_column(parameter)
-            elif isinstance(parameter, list):
-                # for list input, we are assuming only one element type exist in the list.
-                # for empty list, we are converting it into an empty long[] in the JVM side.
-                gateway = self._sc._gateway
-                assert gateway is not None
-                jclass = gateway.jvm.long
-                if len(parameter) >= 1:
-                    mapping = {
-                        str: gateway.jvm.java.lang.String,
-                        float: gateway.jvm.double,
-                        int: gateway.jvm.long,
-                    }
-                    jclass = mapping[type(parameter[0])]
-                return toJArray(gateway, jclass, parameter)
-            else:
-                return parameter
-
-        jdf = self._jdf.hint(name, self._jseq(parameters, _converter))
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def count(self) -> int:
         """Returns the number of rows in this :class:`DataFrame`.
 
@@ -1392,8 +1176,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.count()
         3
         """
-        return int(self._jdf.count())
+        ...
 
+    @dispatch_df_method
     def collect(self) -> List[Row]:
         """Returns all the records in the DataFrame as a list of :class:`Row`.
 
@@ -1459,10 +1244,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> [row.asDict() for row in rows]
         [{'age': 14, 'name': 'Tom'}, {'age': 23, 'name': 'Alice'}, {'age': 16, 'name': 'Bob'}]
         """
-        with SCCallSiteSync(self._sc):
-            sock_info = self._jdf.collectToPython()
-        return list(_load_from_socket(sock_info, BatchedSerializer(CPickleSerializer())))
+        ...
 
+    @dispatch_df_method
     def toLocalIterator(self, prefetchPartitions: bool = False) -> Iterator[Row]:
         """
         Returns an iterator that contains all of the rows in this :class:`DataFrame`.
@@ -1495,10 +1279,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> list(df.toLocalIterator())
         [Row(age=14, name='Tom'), Row(age=23, name='Alice'), Row(age=16, name='Bob')]
         """
-        with SCCallSiteSync(self._sc):
-            sock_info = self._jdf.toPythonIterator(prefetchPartitions)
-        return _local_iterator_from_socket(sock_info, BatchedSerializer(CPickleSerializer()))
+        ...
 
+    @dispatch_df_method
     def limit(self, num: int) -> "DataFrame":
         """Limits the result count to the number specified.
 
@@ -1534,9 +1317,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         +---+----+
         +---+----+
         """
-        jdf = self._jdf.limit(num)
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def offset(self, num: int) -> "DataFrame":
         """Returns a new :class: `DataFrame` by skipping the first `n` rows.
 
@@ -1572,9 +1355,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         +---+----+
         +---+----+
         """
-        jdf = self._jdf.offset(num)
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def take(self, num: int) -> List[Row]:
         """Returns the first ``num`` rows as a :class:`list` of :class:`Row`.
 
@@ -1604,8 +1387,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.take(2)
         [Row(age=14, name='Tom'), Row(age=23, name='Alice')]
         """
-        return self.limit(num).collect()
+        ...
 
+    @dispatch_df_method
     def tail(self, num: int) -> List[Row]:
         """
         Returns the last ``num`` rows as a :class:`list` of :class:`Row`.
@@ -1637,10 +1421,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.tail(2)
         [Row(age=23, name='Alice'), Row(age=16, name='Bob')]
         """
-        with SCCallSiteSync(self._sc):
-            sock_info = self._jdf.tailToPython(num)
-        return list(_load_from_socket(sock_info, BatchedSerializer(CPickleSerializer())))
+        ...
 
+    @dispatch_df_method
     def foreach(self, f: Callable[[Row], None]) -> None:
         """Applies the ``f`` function to all :class:`Row` of this :class:`DataFrame`.
 
@@ -1666,8 +1449,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...
         >>> df.foreach(func)
         """
-        self.rdd.foreach(f)
+        ...
 
+    @dispatch_df_method
     def foreachPartition(self, f: Callable[[Iterator[Row]], None]) -> None:
         """Applies the ``f`` function to each partition of this :class:`DataFrame`.
 
@@ -1694,8 +1478,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...
         >>> df.foreachPartition(func)
         """
-        self.rdd.foreachPartition(f)  # type: ignore[arg-type]
+        ...
 
+    @dispatch_df_method
     def cache(self) -> "DataFrame":
         """Persists the :class:`DataFrame` with the default storage level (`MEMORY_AND_DISK_DESER`).
 
@@ -1723,10 +1508,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         == Physical Plan ==
         InMemoryTableScan ...
         """
-        self.is_cached = True
-        self._jdf.cache()
-        return self
+        ...
 
+    @dispatch_df_method
     def persist(
         self,
         storageLevel: StorageLevel = (StorageLevel.MEMORY_AND_DISK_DESER),
@@ -1771,10 +1555,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.persist(StorageLevel.DISK_ONLY)
         DataFrame[id: bigint]
         """
-        self.is_cached = True
-        javaStorageLevel = self._sc._getJavaStorageLevel(storageLevel)
-        self._jdf.persist(javaStorageLevel)
-        return self
+        ...
 
     @property
     def storageLevel(self) -> StorageLevel:
@@ -1802,16 +1583,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df2.persist(StorageLevel.DISK_ONLY_2).storageLevel
         StorageLevel(True, False, False, False, 2)
         """
-        java_storage_level = self._jdf.storageLevel()
-        storage_level = StorageLevel(
-            java_storage_level.useDisk(),
-            java_storage_level.useMemory(),
-            java_storage_level.useOffHeap(),
-            java_storage_level.deserialized(),
-            java_storage_level.replication(),
-        )
-        return storage_level
+        ...
 
+    @dispatch_df_method
     def unpersist(self, blocking: bool = False) -> "DataFrame":
         """Marks the :class:`DataFrame` as non-persistent, and remove all blocks for it from
         memory and disk.
@@ -1850,6 +1624,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         self._jdf.unpersist(blocking)
         return self
 
+    @dispatch_df_method
     def coalesce(self, numPartitions: int) -> "DataFrame":
         """
         Returns a new :class:`DataFrame` that has exactly `numPartitions` partitions.
@@ -1915,7 +1690,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def repartition(self, *cols: "ColumnOrName") -> "DataFrame":
         ...
 
-    def repartition(  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def repartition(
         self, numPartitions: Union[int, "ColumnOrName"], *cols: "ColumnOrName"
     ) -> "DataFrame":
         """
@@ -2020,25 +1796,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |        2|
         +---------+
         """
-        if isinstance(numPartitions, int):
-            if len(cols) == 0:
-                return DataFrame(self._jdf.repartition(numPartitions), self.sparkSession)
-            else:
-                return DataFrame(
-                    self._jdf.repartition(numPartitions, self._jcols(*cols)),
-                    self.sparkSession,
-                )
-        elif isinstance(numPartitions, (str, Column)):
-            cols = (numPartitions,) + cols
-            return DataFrame(self._jdf.repartition(self._jcols(*cols)), self.sparkSession)
-        else:
-            raise PySparkTypeError(
-                error_class="NOT_COLUMN_OR_STR",
-                message_parameters={
-                    "arg_name": "numPartitions",
-                    "arg_type": type(numPartitions).__name__,
-                },
-            )
+        ...
 
     @overload
     def repartitionByRange(self, numPartitions: int, *cols: "ColumnOrName") -> "DataFrame":
@@ -2048,7 +1806,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def repartitionByRange(self, *cols: "ColumnOrName") -> "DataFrame":
         ...
 
-    def repartitionByRange(  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def repartitionByRange(
         self, numPartitions: Union[int, "ColumnOrName"], *cols: "ColumnOrName"
     ) -> "DataFrame":
         """
@@ -2104,29 +1863,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         | 23|Alice|                   1|
         +---+-----+--------------------+
         """
-        if isinstance(numPartitions, int):
-            if len(cols) == 0:
-                raise PySparkValueError(
-                    error_class="CANNOT_BE_EMPTY",
-                    message_parameters={"item": "partition-by expression"},
-                )
-            else:
-                return DataFrame(
-                    self._jdf.repartitionByRange(numPartitions, self._jcols(*cols)),
-                    self.sparkSession,
-                )
-        elif isinstance(numPartitions, (str, Column)):
-            cols = (numPartitions,) + cols
-            return DataFrame(self._jdf.repartitionByRange(self._jcols(*cols)), self.sparkSession)
-        else:
-            raise PySparkTypeError(
-                error_class="NOT_COLUMN_OR_INT_OR_STR",
-                message_parameters={
-                    "arg_name": "numPartitions",
-                    "arg_type": type(numPartitions).__name__,
-                },
-            )
+        ...
 
+    @dispatch_df_method
     def distinct(self) -> "DataFrame":
         """Returns a new :class:`DataFrame` containing the distinct rows in this :class:`DataFrame`.
 
@@ -2225,7 +1964,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         | 23|Alice|     F|
         +---+-----+------+
         """
-        return DataFrame(self._jdf.distinct(), self.sparkSession)
+        ...
 
     @overload
     def sample(self, fraction: float, seed: Optional[int] = ...) -> "DataFrame":
@@ -2240,7 +1979,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     ) -> "DataFrame":
         ...
 
-    def sample(  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def sample(
         self,
         withReplacement: Optional[Union[float, bool]] = None,
         fraction: Optional[Union[int, float]] = None,
@@ -2290,47 +2030,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.sample(False, fraction=1.0).count()
         10
         """
+        ...
 
-        # For the cases below:
-        #   sample(True, 0.5 [, seed])
-        #   sample(True, fraction=0.5 [, seed])
-        #   sample(withReplacement=False, fraction=0.5 [, seed])
-        is_withReplacement_set = type(withReplacement) == bool and isinstance(fraction, float)
-
-        # For the case below:
-        #   sample(faction=0.5 [, seed])
-        is_withReplacement_omitted_kwargs = withReplacement is None and isinstance(fraction, float)
-
-        # For the case below:
-        #   sample(0.5 [, seed])
-        is_withReplacement_omitted_args = isinstance(withReplacement, float)
-
-        if not (
-            is_withReplacement_set
-            or is_withReplacement_omitted_kwargs
-            or is_withReplacement_omitted_args
-        ):
-            argtypes = [type(arg).__name__ for arg in [withReplacement, fraction, seed]]
-            raise PySparkTypeError(
-                error_class="NOT_BOOL_OR_FLOAT_OR_INT",
-                message_parameters={
-                    "arg_name": "withReplacement (optional), "
-                    + "fraction (required) and seed (optional)",
-                    "arg_type": ", ".join(argtypes),
-                },
-            )
-
-        if is_withReplacement_omitted_args:
-            if fraction is not None:
-                seed = cast(int, fraction)
-            fraction = withReplacement
-            withReplacement = None
-
-        seed = int(seed) if seed is not None else None
-        args = [arg for arg in [withReplacement, fraction, seed] if arg is not None]
-        jdf = self._jdf.sample(*args)
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def sampleBy(
         self, col: "ColumnOrName", fractions: Dict[Any, float], seed: Optional[int] = None
     ) -> "DataFrame":
@@ -2375,36 +2077,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> dataset.sampleBy(col("key"), fractions={2: 1.0}, seed=0).count()
         33
         """
-        if isinstance(col, str):
-            col = Column(col)
-        elif not isinstance(col, Column):
-            raise PySparkTypeError(
-                error_class="NOT_COLUMN_OR_STR",
-                message_parameters={"arg_name": "col", "arg_type": type(col).__name__},
-            )
-        if not isinstance(fractions, dict):
-            raise PySparkTypeError(
-                error_class="NOT_DICT",
-                message_parameters={"arg_name": "fractions", "arg_type": type(fractions).__name__},
-            )
-        for k, v in fractions.items():
-            if not isinstance(k, (float, int, str)):
-                raise PySparkTypeError(
-                    error_class="DISALLOWED_TYPE_FOR_CONTAINER",
-                    message_parameters={
-                        "arg_name": "fractions",
-                        "arg_type": type(fractions).__name__,
-                        "allowed_types": "float, int, str",
-                        "item_type": type(k).__name__,
-                    },
-                )
-            fractions[k] = float(v)
-        col = col._jc
-        seed = seed if seed is not None else random.randint(0, sys.maxsize)
-        return DataFrame(
-            self._jdf.stat().sampleBy(col, self._jmap(fractions), seed), self.sparkSession
-        )
+        ...
 
+    @dispatch_df_method
     def randomSplit(self, weights: List[float], seed: Optional[int] = None) -> List["DataFrame"]:
         """Randomly splits this :class:`DataFrame` with the provided weights.
 
@@ -2442,17 +2117,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> splits[1].count()
         2
         """
-        for w in weights:
-            if w < 0.0:
-                raise PySparkValueError(
-                    error_class="VALUE_NOT_POSITIVE",
-                    message_parameters={"arg_name": "weights", "arg_value": str(w)},
-                )
-        seed = seed if seed is not None else random.randint(0, sys.maxsize)
-        df_array = self._jdf.randomSplit(
-            _to_list(self.sparkSession._sc, cast(List["ColumnOrName"], weights)), int(seed)
-        )
-        return [DataFrame(df, self.sparkSession) for df in df_array]
+        ...
 
     @property
     def dtypes(self) -> List[Tuple[str, str]]:
@@ -2475,7 +2140,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.dtypes
         [('age', 'bigint'), ('name', 'string')]
         """
-        return [(str(f.name), f.dataType.simpleString()) for f in self.schema.fields]
+        ...
 
     @property
     def columns(self) -> List[str]:
@@ -2552,8 +2217,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.columns == df2.columns
         False
         """
-        return [f.name for f in self.schema.fields]
+        ...
 
+    @dispatch_df_method
     def colRegex(self, colName: str) -> Column:
         """
         Selects column based on the column name specified as a regex and returns it
@@ -2585,14 +2251,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |   3|
         +----+
         """
-        if not isinstance(colName, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "colName", "arg_type": type(colName).__name__},
-            )
-        jc = self._jdf.colRegex(colName)
-        return Column(jc)
+        ...
 
+    @dispatch_df_method
     def to(self, schema: StructType) -> "DataFrame":
         """
         Returns a new :class:`DataFrame` where each row is reconciled to match the specified
@@ -2648,10 +2309,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  1|  a|
         +---+---+
         """
-        assert schema is not None
-        jschema = self._jdf.sparkSession().parseDataType(schema.json())
-        return DataFrame(self._jdf.to(jschema), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def alias(self, alias: str) -> "DataFrame":
         """Returns a new :class:`DataFrame` with an alias set.
 
@@ -2688,9 +2348,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |Alice|Alice| 23|
         +-----+-----+---+
         """
-        assert isinstance(alias, str), "alias should be a string"
-        return DataFrame(getattr(self._jdf, "as")(alias), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def crossJoin(self, other: "DataFrame") -> "DataFrame":
         """Returns the cartesian product with another :class:`DataFrame`.
 
@@ -2728,10 +2388,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         | 16|  Bob|    85|
         +---+-----+------+
         """
+        ...
 
-        jdf = self._jdf.crossJoin(other._jdf)
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def join(
         self,
         other: "DataFrame",
@@ -2911,30 +2570,10 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |Alice|  2|
         +-----+---+
         """
-
-        if on is not None and not isinstance(on, list):
-            on = [on]  # type: ignore[assignment]
-
-        if on is not None:
-            if isinstance(on[0], str):
-                on = self._jseq(cast(List[str], on))
-            else:
-                assert isinstance(on[0], Column), "on should be Column or list of Column"
-                on = reduce(lambda x, y: x.__and__(y), cast(List[Column], on))
-                on = on._jc
-
-        if on is None and how is None:
-            jdf = self._jdf.join(other._jdf)
-        else:
-            if how is None:
-                how = "inner"
-            if on is None:
-                on = self._jseq([])
-            assert isinstance(how, str), "how should be a string"
-            jdf = self._jdf.join(other._jdf, on, how)
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
     # TODO(SPARK-22947): Fix the DataFrame API.
+    @dispatch_df_method
     def _joinAsOf(
         self,
         other: "DataFrame",
@@ -3018,44 +2657,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         [Row(a=1, left_val='a', right_val=1),
          Row(a=5, left_val='b', right_val=6)]
         """
-        if isinstance(leftAsOfColumn, str):
-            leftAsOfColumn = self[leftAsOfColumn]
-        left_as_of_jcol = leftAsOfColumn._jc
-        if isinstance(rightAsOfColumn, str):
-            rightAsOfColumn = other[rightAsOfColumn]
-        right_as_of_jcol = rightAsOfColumn._jc
+        ...
 
-        if on is not None and not isinstance(on, list):
-            on = [on]  # type: ignore[assignment]
-
-        if on is not None:
-            if isinstance(on[0], str):
-                on = self._jseq(cast(List[str], on))
-            else:
-                assert isinstance(on[0], Column), "on should be Column or list of Column"
-                on = reduce(lambda x, y: x.__and__(y), cast(List[Column], on))
-                on = on._jc
-
-        if how is None:
-            how = "inner"
-        assert isinstance(how, str), "how should be a string"
-
-        if tolerance is not None:
-            assert isinstance(tolerance, Column), "tolerance should be Column"
-            tolerance = tolerance._jc
-
-        jdf = self._jdf.joinAsOf(
-            other._jdf,
-            left_as_of_jcol,
-            right_as_of_jcol,
-            on,
-            how,
-            tolerance,
-            allowExactMatches,
-            direction,
-        )
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def sortWithinPartitions(
         self,
         *cols: Union[int, str, Column, List[Union[int, str, Column]]],
@@ -3117,9 +2721,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  2|Alice|
         +---+-----+
         """
-        jdf = self._jdf.sortWithinPartitions(self._sort_cols(cols, kwargs))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def sort(
         self,
         *cols: Union[int, str, Column, List[Union[int, str, Column]]],
@@ -3277,98 +2881,11 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  2|Alice|
         +---+-----+
         """
-        jdf = self._jdf.sort(self._sort_cols(cols, kwargs))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
     orderBy = sort
 
-    def _jseq(
-        self,
-        cols: Sequence,
-        converter: Optional[Callable[..., Union["PrimitiveType", "JavaObject"]]] = None,
-    ) -> "JavaObject":
-        """Return a JVM Seq of Columns from a list of Column or names"""
-        return _to_seq(self.sparkSession._sc, cols, converter)
-
-    def _jmap(self, jm: Dict) -> "JavaObject":
-        """Return a JVM Scala Map from a dict"""
-        return _to_scala_map(self.sparkSession._sc, jm)
-
-    def _jcols(self, *cols: "ColumnOrName") -> "JavaObject":
-        """Return a JVM Seq of Columns from a list of Column or column names
-
-        If `cols` has only one list in it, cols[0] will be used as the list.
-        """
-        if len(cols) == 1 and isinstance(cols[0], list):
-            cols = cols[0]
-        return self._jseq(cols, _to_java_column)
-
-    def _jcols_ordinal(self, *cols: "ColumnOrNameOrOrdinal") -> "JavaObject":
-        """Return a JVM Seq of Columns from a list of Column or column names or column ordinals.
-
-        If `cols` has only one list in it, cols[0] will be used as the list.
-        """
-        if len(cols) == 1 and isinstance(cols[0], list):
-            cols = cols[0]
-
-        _cols = []
-        for c in cols:
-            if isinstance(c, int) and not isinstance(c, bool):
-                if c < 1:
-                    raise PySparkIndexError(
-                        error_class="INDEX_NOT_POSITIVE", message_parameters={"index": str(c)}
-                    )
-                # ordinal is 1-based
-                _cols.append(self[c - 1])
-            else:
-                _cols.append(c)  # type: ignore[arg-type]
-        return self._jseq(_cols, _to_java_column)
-
-    def _sort_cols(
-        self,
-        cols: Sequence[Union[int, str, Column, List[Union[int, str, Column]]]],
-        kwargs: Dict[str, Any],
-    ) -> "JavaObject":
-        """Return a JVM Seq of Columns that describes the sort order"""
-        if not cols:
-            raise PySparkValueError(
-                error_class="CANNOT_BE_EMPTY",
-                message_parameters={"item": "column"},
-            )
-        if len(cols) == 1 and isinstance(cols[0], list):
-            cols = cols[0]
-
-        jcols = []
-        for c in cols:
-            if isinstance(c, int) and not isinstance(c, bool):
-                # ordinal is 1-based
-                if c > 0:
-                    _c = self[c - 1]
-                # negative ordinal means sort by desc
-                elif c < 0:
-                    _c = self[-c - 1].desc()
-                else:
-                    raise PySparkIndexError(
-                        error_class="ZERO_INDEX",
-                        message_parameters={},
-                    )
-            else:
-                _c = c  # type: ignore[assignment]
-            jcols.append(_to_java_column(cast("ColumnOrName", _c)))
-
-        ascending = kwargs.get("ascending", True)
-        if isinstance(ascending, (bool, int)):
-            if not ascending:
-                jcols = [jc.desc() for jc in jcols]
-        elif isinstance(ascending, list):
-            jcols = [jc if asc else jc.desc() for asc, jc in zip(ascending, jcols)]
-        else:
-            raise PySparkTypeError(
-                error_class="NOT_BOOL_OR_LIST",
-                message_parameters={"arg_name": "ascending", "arg_type": type(ascending).__name__},
-            )
-        return self._jseq(jcols)
-
+    @dispatch_df_method
     def describe(self, *cols: Union[str, List[str]]) -> "DataFrame":
         """Computes basic statistics for numeric and string columns.
 
@@ -3430,11 +2947,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         --------
         DataFrame.summary
         """
-        if len(cols) == 1 and isinstance(cols[0], list):
-            cols = cols[0]  # type: ignore[assignment]
-        jdf = self._jdf.describe(self._jseq(cols))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def summary(self, *statistics: str) -> "DataFrame":
         """Computes specified statistics for numeric and string columns. Available statistics are:
         - count
@@ -3503,10 +3018,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         --------
         DataFrame.display
         """
-        if len(statistics) == 1 and isinstance(statistics[0], list):
-            statistics = statistics[0]
-        jdf = self._jdf.summary(self._jseq(statistics))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
     @overload
     def head(self) -> Optional[Row]:
@@ -3516,6 +3028,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def head(self, n: int) -> List[Row]:
         ...
 
+    @dispatch_df_method
     def head(self, n: Optional[int] = None) -> Union[Optional[Row], List[Row]]:
         """Returns the first ``n`` rows.
 
@@ -3551,11 +3064,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.head(0)
         []
         """
-        if n is None:
-            rs = self.head(1)
-            return rs[0] if rs else None
-        return self.take(n)
+        ...
 
+    @dispatch_df_method
     def first(self) -> Optional[Row]:
         """Returns the first row as a :class:`Row`.
 
@@ -3576,7 +3087,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df.first()
         Row(age=2, name='Alice')
         """
-        return self.head()
+        ...
 
     @overload
     def __getitem__(self, item: Union[int, str]) -> Column:
@@ -3586,6 +3097,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def __getitem__(self, item: Union[Column, List, Tuple]) -> "DataFrame":
         ...
 
+    @dispatch_df_method
     def __getitem__(self, item: Union[int, str, Column, List, Tuple]) -> Union[Column, "DataFrame"]:
         """Returns the column as a :class:`Column`.
 
@@ -3658,22 +3170,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5| Bob|
         +---+----+
         """
-        if isinstance(item, str):
-            jc = self._jdf.apply(item)
-            return Column(jc)
-        elif isinstance(item, Column):
-            return self.filter(item)
-        elif isinstance(item, (list, tuple)):
-            return self.select(*item)
-        elif isinstance(item, int):
-            jc = self._jdf.apply(self.columns[item])
-            return Column(jc)
-        else:
-            raise PySparkTypeError(
-                error_class="NOT_COLUMN_OR_FLOAT_OR_INT_OR_LIST_OR_STR",
-                message_parameters={"arg_name": "item", "arg_type": type(item).__name__},
-            )
+        ...
 
+    @dispatch_df_method
     def __getattr__(self, name: str) -> Column:
         """Returns the :class:`Column` denoted by ``name``.
 
@@ -3707,13 +3206,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5|
         +---+
         """
-        if name not in self.columns:
-            raise PySparkAttributeError(
-                error_class="ATTRIBUTE_NOT_SUPPORTED", message_parameters={"attr_name": name}
-            )
-        jc = self._jdf.apply(name)
-        return Column(jc)
+        ...
 
+    @dispatch_df_method
     def __dir__(self) -> List[str]:
         """
         Examples
@@ -3751,9 +3246,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> [attr for attr in dir(df) if attr[0] == 'i'][:7] # Doesn't include 1 or name 1
         ['i_like_pancakes', 'id', 'id2', 'inputFiles', 'intersect', 'intersectAll', 'isEmpty']
         """
-        attrs = set(super().__dir__())
-        attrs.update(filter(lambda s: s.isidentifier(), self.columns))
-        return sorted(attrs)
+        ...
 
     @overload
     def select(self, *cols: "ColumnOrName") -> "DataFrame":
@@ -3763,7 +3256,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def select(self, __cols: Union[List[Column], List[str]]) -> "DataFrame":
         ...
 
-    def select(self, *cols: "ColumnOrName") -> "DataFrame":  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def select(self, *cols: "ColumnOrName") -> "DataFrame":
         """Projects a set of expressions and returns a new :class:`DataFrame`.
 
         .. versionadded:: 1.3.0
@@ -3808,8 +3302,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  Bob| 15|
         +-----+---+
         """
-        jdf = self._jdf.select(self._jcols(*cols))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
     @overload
     def selectExpr(self, *expr: str) -> "DataFrame":
@@ -3819,6 +3312,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def selectExpr(self, *expr: List[str]) -> "DataFrame":
         ...
 
+    @dispatch_df_method
     def selectExpr(self, *expr: Union[str, List[str]]) -> "DataFrame":
         """Projects a set of SQL expressions and returns a new :class:`DataFrame`.
 
@@ -3846,11 +3340,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |       10|       5|
         +---------+--------+
         """
-        if len(expr) == 1 and isinstance(expr[0], list):
-            expr = expr[0]  # type: ignore[assignment]
-        jdf = self._jdf.selectExpr(self._jseq(expr))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def filter(self, condition: "ColumnOrName") -> "DataFrame":
         """Filters rows using the given condition.
 
@@ -4004,16 +3496,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5|  Bob|Physics|
         +---+-----+-------+
         """
-        if isinstance(condition, str):
-            jdf = self._jdf.filter(condition)
-        elif isinstance(condition, Column):
-            jdf = self._jdf.filter(condition._jc)
-        else:
-            raise PySparkTypeError(
-                error_class="NOT_COLUMN_OR_STR",
-                message_parameters={"arg_name": "condition", "arg_type": type(condition).__name__},
-            )
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
     @overload
     def groupBy(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":
@@ -4023,7 +3506,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def groupBy(self, __cols: Union[List[Column], List[str], List[int]]) -> "GroupedData":
         ...
 
-    def groupBy(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def groupBy(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":
         """
         Groups the :class:`DataFrame` by the specified columns so that aggregation
         can be performed on them.
@@ -4122,10 +3606,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  Bob|  5|    1|
         +-----+---+-----+
         """
-        jgd = self._jdf.groupBy(self._jcols_ordinal(*cols))
-        from pyspark.sql.group import GroupedData
-
-        return GroupedData(jgd, self)
+        ...
 
     @overload
     def rollup(self, *cols: "ColumnOrName") -> "GroupedData":
@@ -4135,7 +3616,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def rollup(self, __cols: Union[List[Column], List[str]]) -> "GroupedData":
         ...
 
-    def rollup(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def rollup(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":
         """
         Create a multi-dimensional rollup for the current :class:`DataFrame` using
         the specified columns, allowing for aggregation on them.
@@ -4207,10 +3689,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  Bob|   5|    1|
         +-----+----+-----+
         """
-        jgd = self._jdf.rollup(self._jcols_ordinal(*cols))
-        from pyspark.sql.group import GroupedData
-
-        return GroupedData(jgd, self)
+        ...
 
     @overload
     def cube(self, *cols: "ColumnOrName") -> "GroupedData":
@@ -4220,7 +3699,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def cube(self, __cols: Union[List[Column], List[str]]) -> "GroupedData":
         ...
 
-    def cube(self, *cols: "ColumnOrName") -> "GroupedData":  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def cube(self, *cols: "ColumnOrName") -> "GroupedData":
         """
         Create a multi-dimensional cube for the current :class:`DataFrame` using
         the specified columns, allowing aggregations to be performed on them.
@@ -4297,11 +3777,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  Bob|   5|    1|
         +-----+----+-----+
         """
-        jgd = self._jdf.cube(self._jcols_ordinal(*cols))
-        from pyspark.sql.group import GroupedData
+        ...
 
-        return GroupedData(jgd, self)
-
+    @dispatch_df_method
     def groupingSets(
         self, groupingSets: Sequence[Sequence["ColumnOrName"]], *cols: "ColumnOrName"
     ) -> "GroupedData":
@@ -4391,13 +3869,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         --------
         GroupedData
         """
-        from pyspark.sql.group import GroupedData
+        ...
 
-        jgrouping_sets = _to_seq(self._sc, [self._jcols(*inner) for inner in groupingSets])
-
-        jgd = self._jdf.groupingSets(jgrouping_sets, self._jcols(*cols))
-        return GroupedData(jgd, self)
-
+    @dispatch_df_method
     def unpivot(
         self,
         ids: Union["ColumnOrName", List["ColumnOrName"], Tuple["ColumnOrName", ...]],
@@ -4479,26 +3953,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         --------
         DataFrame.melt
         """
-        assert ids is not None, "ids must not be None"
+        ...
 
-        def to_jcols(
-            cols: Union["ColumnOrName", List["ColumnOrName"], Tuple["ColumnOrName", ...]]
-        ) -> "JavaObject":
-            if isinstance(cols, list):
-                return self._jcols(*cols)
-            if isinstance(cols, tuple):
-                return self._jcols(*list(cols))
-            return self._jcols(cols)
-
-        jids = to_jcols(ids)
-        if values is None:
-            jdf = self._jdf.unpivotWithSeq(jids, variableColumnName, valueColumnName)
-        else:
-            jvals = to_jcols(values)
-            jdf = self._jdf.unpivotWithSeq(jids, jvals, variableColumnName, valueColumnName)
-
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def melt(
         self,
         ids: Union["ColumnOrName", List["ColumnOrName"], Tuple["ColumnOrName", ...]],
@@ -4542,8 +3999,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         -----
         Supports Spark Connect.
         """
-        return self.unpivot(ids, values, variableColumnName, valueColumnName)
+        ...
 
+    @dispatch_df_method
     def agg(self, *exprs: Union[Column, Dict[str, str]]) -> "DataFrame":
         """Aggregate on the entire :class:`DataFrame` without groups
         (shorthand for ``df.groupBy().agg()``).
@@ -4580,8 +4038,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |       2|
         +--------+
         """
-        return self.groupBy().agg(*exprs)  # type: ignore[arg-type]
+        ...
 
+    @dispatch_df_method
     def observe(
         self,
         observation: Union["Observation", str],
@@ -4679,37 +4138,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...     count(col("error")).alias("erc"))  # doctest: +SKIP
         >>> observed_ds.writeStream.format("console").start()  # doctest: +SKIP
         """
-        from pyspark.sql import Observation
+        ...
 
-        if len(exprs) == 0:
-            raise PySparkValueError(
-                error_class="CANNOT_BE_EMPTY",
-                message_parameters={"item": "exprs"},
-            )
-        if not all(isinstance(c, Column) for c in exprs):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OF_COLUMN",
-                message_parameters={"arg_name": "exprs"},
-            )
-
-        if isinstance(observation, Observation):
-            return observation._on(self, *exprs)
-        elif isinstance(observation, str):
-            return DataFrame(
-                self._jdf.observe(
-                    observation, exprs[0]._jc, _to_seq(self._sc, [c._jc for c in exprs[1:]])
-                ),
-                self.sparkSession,
-            )
-        else:
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OF_COLUMN",
-                message_parameters={
-                    "arg_name": "observation",
-                    "arg_type": type(observation).__name__,
-                },
-            )
-
+    @dispatch_df_method
     def union(self, other: "DataFrame") -> "DataFrame":
         """Return a new :class:`DataFrame` containing the union of rows in this and another
         :class:`DataFrame`.
@@ -4806,8 +4237,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  4|    D|
         +---+-----+
         """
-        return DataFrame(self._jdf.union(other._jdf), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def unionAll(self, other: "DataFrame") -> "DataFrame":
         """Return a new :class:`DataFrame` containing the union of rows in this and another
         :class:`DataFrame`.
@@ -4840,8 +4272,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         --------
         DataFrame.union
         """
-        return self.union(other)
+        ...
 
+    @dispatch_df_method
     def unionByName(self, other: "DataFrame", allowMissingColumns: bool = False) -> "DataFrame":
         """Returns a new :class:`DataFrame` containing union of rows in this and another
         :class:`DataFrame`.
@@ -4920,8 +4353,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |NULL|NULL|NULL|   3|   4|   5|
         +----+----+----+----+----+----+
         """
-        return DataFrame(self._jdf.unionByName(other._jdf, allowMissingColumns), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def intersect(self, other: "DataFrame") -> "DataFrame":
         """Return a new :class:`DataFrame` containing rows only in
         both this :class:`DataFrame` and another :class:`DataFrame`.
@@ -4986,8 +4420,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  1|  2|
         +---+---+
         """
-        return DataFrame(self._jdf.intersect(other._jdf), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def intersectAll(self, other: "DataFrame") -> "DataFrame":
         """Return a new :class:`DataFrame` containing rows in both this :class:`DataFrame`
         and another :class:`DataFrame` while preserving duplicates.
@@ -5051,8 +4486,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  1|  2|
         +---+---+
         """
-        return DataFrame(self._jdf.intersectAll(other._jdf), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def subtract(self, other: "DataFrame") -> "DataFrame":
         """Return a new :class:`DataFrame` containing rows in this :class:`DataFrame`
         but not in another :class:`DataFrame`.
@@ -5113,8 +4549,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         +---+---+
         +---+---+
         """
-        return DataFrame(getattr(self._jdf, "except")(other._jdf), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def dropDuplicates(self, subset: Optional[List[str]] = None) -> "DataFrame":
         """Return a new :class:`DataFrame` with duplicate rows removed,
         optionally only considering certain columns.
@@ -5170,18 +4607,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |Alice|  5|    80|
         +-----+---+------+
         """
-        if subset is not None and (not isinstance(subset, Iterable) or isinstance(subset, str)):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_TUPLE",
-                message_parameters={"arg_name": "subset", "arg_type": type(subset).__name__},
-            )
+        ...
 
-        if subset is None:
-            jdf = self._jdf.dropDuplicates()
-        else:
-            jdf = self._jdf.dropDuplicates(self._jseq(subset))
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def dropDuplicatesWithinWatermark(self, subset: Optional[List[str]] = None) -> "DataFrame":
         """Return a new :class:`DataFrame` with duplicate rows removed,
          optionally only considering certain columns, within watermark.
@@ -5230,18 +4658,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
 
          >>> df.dropDuplicatesWithinWatermark(['value'])  # doctest: +SKIP
         """
-        if subset is not None and (not isinstance(subset, Iterable) or isinstance(subset, str)):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_TUPLE",
-                message_parameters={"arg_name": "subset", "arg_type": type(subset).__name__},
-            )
+        ...
 
-        if subset is None:
-            jdf = self._jdf.dropDuplicatesWithinWatermark()
-        else:
-            jdf = self._jdf.dropDuplicatesWithinWatermark(self._jseq(subset))
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def dropna(
         self,
         how: str = "any",
@@ -5323,26 +4742,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5|  NULL|  Bob|
         +---+------+-----+
         """
-        if how is not None and how not in ["any", "all"]:
-            raise PySparkValueError(
-                error_class="VALUE_NOT_ANY_OR_ALL",
-                message_parameters={"arg_name": "how", "arg_type": how},
-            )
-
-        if subset is None:
-            subset = self.columns
-        elif isinstance(subset, str):
-            subset = [subset]
-        elif not isinstance(subset, (list, tuple)):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_STR_OR_TUPLE",
-                message_parameters={"arg_name": "subset", "arg_type": type(subset).__name__},
-            )
-
-        if thresh is None:
-            thresh = len(subset) if how == "any" else 1
-
-        return DataFrame(self._jdf.na().drop(thresh, self._jseq(subset)), self.sparkSession)
+        ...
 
     @overload
     def fillna(
@@ -5356,6 +4756,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def fillna(self, value: Dict[str, "LiteralType"]) -> "DataFrame":
         ...
 
+    @dispatch_df_method
     def fillna(
         self,
         value: Union["LiteralType", Dict[str, "LiteralType"]],
@@ -5445,32 +4846,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |NULL|  NULL|Spark|true|
         +----+------+-----+----+
         """
-        if not isinstance(value, (float, int, str, bool, dict)):
-            raise PySparkTypeError(
-                error_class="NOT_BOOL_OR_DICT_OR_FLOAT_OR_INT_OR_STR",
-                message_parameters={"arg_name": "value", "arg_type": type(value).__name__},
-            )
-
-        # Note that bool validates isinstance(int), but we don't want to
-        # convert bools to floats
-
-        if not isinstance(value, bool) and isinstance(value, int):
-            value = float(value)
-
-        if isinstance(value, dict):
-            return DataFrame(self._jdf.na().fill(value), self.sparkSession)
-        elif subset is None:
-            return DataFrame(self._jdf.na().fill(value), self.sparkSession)
-        else:
-            if isinstance(subset, str):
-                subset = [subset]
-            elif not isinstance(subset, (list, tuple)):
-                raise PySparkTypeError(
-                    error_class="NOT_LIST_OR_TUPLE",
-                    message_parameters={"arg_name": "subset", "arg_type": type(subset).__name__},
-                )
-
-            return DataFrame(self._jdf.na().fill(value, self._jseq(subset)), self.sparkSession)
+        ...
 
     @overload
     def replace(
@@ -5507,7 +4883,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     ) -> "DataFrame":
         ...
 
-    def replace(  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def replace(
         self,
         to_replace: Union[
             "LiteralType", List["LiteralType"], Dict["LiteralType", "OptionalPrimitiveType"]
@@ -5610,111 +4987,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |NULL|  NULL| NULL|
         +----+------+-----+
         """
-        if value is _NoValue:
-            if isinstance(to_replace, dict):
-                value = None
-            else:
-                raise PySparkTypeError(
-                    error_class="ARGUMENT_REQUIRED",
-                    message_parameters={"arg_name": "value", "condition": "`to_replace` is dict"},
-                )
-
-        # Helper functions
-        def all_of(types: Union[Type, Tuple[Type, ...]]) -> Callable[[Iterable], bool]:
-            """Given a type or tuple of types and a sequence of xs
-            check if each x is instance of type(s)
-
-            >>> all_of(bool)([True, False])
-            True
-            >>> all_of(str)(["a", 1])
-            False
-            """
-
-            def all_of_(xs: Iterable) -> bool:
-                return all(isinstance(x, types) for x in xs)
-
-            return all_of_
-
-        all_of_bool = all_of(bool)
-        all_of_str = all_of(str)
-        all_of_numeric = all_of((float, int))
-
-        # Validate input types
-        valid_types = (bool, float, int, str, list, tuple)
-        if not isinstance(to_replace, valid_types + (dict,)):
-            raise PySparkTypeError(
-                error_class="NOT_BOOL_OR_DICT_OR_FLOAT_OR_INT_OR_LIST_OR_STR_OR_TUPLE",
-                message_parameters={
-                    "arg_name": "to_replace",
-                    "arg_type": type(to_replace).__name__,
-                },
-            )
-
-        if (
-            not isinstance(value, valid_types)
-            and value is not None
-            and not isinstance(to_replace, dict)
-        ):
-            raise PySparkTypeError(
-                error_class="NOT_BOOL_OR_FLOAT_OR_INT_OR_LIST_OR_NONE_OR_STR_OR_TUPLE",
-                message_parameters={
-                    "arg_name": "value",
-                    "arg_type": type(value).__name__,
-                },
-            )
-
-        if isinstance(to_replace, (list, tuple)) and isinstance(value, (list, tuple)):
-            if len(to_replace) != len(value):
-                raise PySparkValueError(
-                    error_class="LENGTH_SHOULD_BE_THE_SAME",
-                    message_parameters={
-                        "arg1": "to_replace",
-                        "arg2": "value",
-                        "arg1_length": str(len(to_replace)),
-                        "arg2_length": str(len(value)),
-                    },
-                )
-
-        if not (subset is None or isinstance(subset, (list, tuple, str))):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_STR_OR_TUPLE",
-                message_parameters={"arg_name": "subset", "arg_type": type(subset).__name__},
-            )
-
-        # Reshape input arguments if necessary
-        if isinstance(to_replace, (float, int, str)):
-            to_replace = [to_replace]
-
-        if isinstance(to_replace, dict):
-            rep_dict = to_replace
-            if value is not None:
-                warnings.warn("to_replace is a dict and value is not None. value will be ignored.")
-        else:
-            if isinstance(value, (float, int, str)) or value is None:
-                value = [value for _ in range(len(to_replace))]
-            rep_dict = dict(zip(to_replace, cast("Iterable[Optional[Union[float, str]]]", value)))
-
-        if isinstance(subset, str):
-            subset = [subset]
-
-        # Verify we were not passed in mixed type generics.
-        if not any(
-            all_of_type(rep_dict.keys())
-            and all_of_type(x for x in rep_dict.values() if x is not None)
-            for all_of_type in [all_of_bool, all_of_str, all_of_numeric]
-        ):
-            raise PySparkValueError(
-                error_class="MIXED_TYPE_REPLACEMENT",
-                message_parameters={},
-            )
-
-        if subset is None:
-            return DataFrame(self._jdf.na().replace("*", rep_dict), self.sparkSession)
-        else:
-            return DataFrame(
-                self._jdf.na().replace(self._jseq(subset), self._jmap(rep_dict)),
-                self.sparkSession,
-            )
+        ...
 
     @overload
     def approxQuantile(
@@ -5734,6 +5007,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     ) -> List[List[float]]:
         ...
 
+    @dispatch_df_method
     def approxQuantile(
         self,
         col: Union[str, List[str], Tuple[str]],
@@ -5830,76 +5104,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> quantiles
         [1.0, 1.0, 5.0]
         """
+        ...
 
-        if not isinstance(col, (str, list, tuple)):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_STR_OR_TUPLE",
-                message_parameters={"arg_name": "col", "arg_type": type(col).__name__},
-            )
-
-        isStr = isinstance(col, str)
-
-        if isinstance(col, tuple):
-            col = list(col)
-        elif isStr:
-            col = [cast(str, col)]
-
-        for c in col:
-            if not isinstance(c, str):
-                raise PySparkTypeError(
-                    error_class="DISALLOWED_TYPE_FOR_CONTAINER",
-                    message_parameters={
-                        "arg_name": "col",
-                        "arg_type": type(col).__name__,
-                        "allowed_types": "str",
-                        "item_type": type(c).__name__,
-                    },
-                )
-        col = _to_list(self._sc, cast(List["ColumnOrName"], col))
-
-        if not isinstance(probabilities, (list, tuple)):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_TUPLE",
-                message_parameters={
-                    "arg_name": "probabilities",
-                    "arg_type": type(probabilities).__name__,
-                },
-            )
-        if isinstance(probabilities, tuple):
-            probabilities = list(probabilities)
-        for p in probabilities:
-            if not isinstance(p, (float, int)) or p < 0 or p > 1:
-                raise PySparkTypeError(
-                    error_class="NOT_LIST_OF_FLOAT_OR_INT",
-                    message_parameters={
-                        "arg_name": "probabilities",
-                        "arg_type": type(p).__name__,
-                    },
-                )
-        probabilities = _to_list(self._sc, cast(List["ColumnOrName"], probabilities))
-
-        if not isinstance(relativeError, (float, int)):
-            raise PySparkTypeError(
-                error_class="NOT_FLOAT_OR_INT",
-                message_parameters={
-                    "arg_name": "relativeError",
-                    "arg_type": type(relativeError).__name__,
-                },
-            )
-        if relativeError < 0:
-            raise PySparkValueError(
-                error_class="NEGATIVE_VALUE",
-                message_parameters={
-                    "arg_name": "relativeError",
-                    "arg_value": str(relativeError),
-                },
-            )
-        relativeError = float(relativeError)
-
-        jaq = self._jdf.stat().approxQuantile(col, probabilities, relativeError)
-        jaq_list = [list(j) for j in jaq]
-        return jaq_list[0] if isStr else jaq_list
-
+    @dispatch_df_method
     def corr(self, col1: str, col2: str, method: Optional[str] = None) -> float:
         """
         Calculates the correlation of two columns of a :class:`DataFrame` as a double value.
@@ -5935,25 +5142,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         1.0
 
         """
-        if not isinstance(col1, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "col1", "arg_type": type(col1).__name__},
-            )
-        if not isinstance(col2, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "col2", "arg_type": type(col2).__name__},
-            )
-        if not method:
-            method = "pearson"
-        if not method == "pearson":
-            raise PySparkValueError(
-                error_class="VALUE_NOT_PEARSON",
-                message_parameters={"arg_name": "method", "arg_value": method},
-            )
-        return self._jdf.stat().corr(col1, col2, method)
+        ...
 
+    @dispatch_df_method
     def cov(self, col1: str, col2: str) -> float:
         """
         Calculate the sample covariance for the given columns, specified by their names, as a
@@ -5986,18 +5177,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         1.0
 
         """
-        if not isinstance(col1, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "col1", "arg_type": type(col1).__name__},
-            )
-        if not isinstance(col2, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "col2", "arg_type": type(col2).__name__},
-            )
-        return self._jdf.stat().cov(col1, col2)
+        ...
 
+    @dispatch_df_method
     def crosstab(self, col1: str, col2: str) -> "DataFrame":
         """
         Computes a pair-wise frequency table of the given columns. Also known as a contingency
@@ -6039,18 +5221,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         +-----+---+---+---+
 
         """
-        if not isinstance(col1, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "col1", "arg_type": type(col1).__name__},
-            )
-        if not isinstance(col2, str):
-            raise PySparkTypeError(
-                error_class="NOT_STR",
-                message_parameters={"arg_name": "col2", "arg_type": type(col2).__name__},
-            )
-        return DataFrame(self._jdf.stat().crosstab(col1, col2), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def freqItems(
         self, cols: Union[List[str], Tuple[str]], support: Optional[float] = None
     ) -> "DataFrame":
@@ -6095,19 +5268,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |   [4, 1, 3]| [8, 11, 10]|
         +------------+------------+
         """
-        if isinstance(cols, tuple):
-            cols = list(cols)
-        if not isinstance(cols, list):
-            raise PySparkTypeError(
-                error_class="NOT_LIST_OR_TUPLE",
-                message_parameters={"arg_name": "cols", "arg_type": type(cols).__name__},
-            )
-        if not support:
-            support = 0.01
-        return DataFrame(
-            self._jdf.stat().freqItems(_to_seq(self._sc, cols), support), self.sparkSession
-        )
+        ...
 
+    @dispatch_df_method
     def _ipython_key_completions_(self) -> List[str]:
         """Returns the names of columns in this :class:`DataFrame`.
 
@@ -6122,8 +5285,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df._ipython_key_completions_()
         ['age 1', 'name?1']
         """
-        return self.columns
+        ...
 
+    @dispatch_df_method
     def withColumns(self, *colsMap: Dict[str, Column]) -> "DataFrame":
         """
         Returns a new :class:`DataFrame` by adding multiple columns or replacing the
@@ -6159,24 +5323,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5|  Bob|   7|   8|
         +---+-----+----+----+
         """
-        # Below code is to help enable kwargs in future.
-        assert len(colsMap) == 1
-        colsMap = colsMap[0]  # type: ignore[assignment]
+        ...
 
-        if not isinstance(colsMap, dict):
-            raise PySparkTypeError(
-                error_class="NOT_DICT",
-                message_parameters={"arg_name": "colsMap", "arg_type": type(colsMap).__name__},
-            )
-
-        col_names = list(colsMap.keys())
-        cols = list(colsMap.values())
-
-        return DataFrame(
-            self._jdf.withColumns(_to_seq(self._sc, col_names), self._jcols(*cols)),
-            self.sparkSession,
-        )
-
+    @dispatch_df_method
     def withColumn(self, colName: str, col: Column) -> "DataFrame":
         """
         Returns a new :class:`DataFrame` by adding a column or replacing the
@@ -6220,13 +5369,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5|  Bob|   7|
         +---+-----+----+
         """
-        if not isinstance(col, Column):
-            raise PySparkTypeError(
-                error_class="NOT_COLUMN",
-                message_parameters={"arg_name": "col", "arg_type": type(col).__name__},
-            )
-        return DataFrame(self._jdf.withColumn(colName, col._jc), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def withColumnRenamed(self, existing: str, new: str) -> "DataFrame":
         """
         Returns a new :class:`DataFrame` by renaming an existing column.
@@ -6287,8 +5432,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |   5|  Bob|
         +----+-----+
         """
-        return DataFrame(self._jdf.withColumnRenamed(existing, new), self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def withColumnsRenamed(self, colsMap: Dict[str, str]) -> "DataFrame":
         """
         Returns a new :class:`DataFrame` by renaming multiple columns.
@@ -6356,25 +5502,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         |  5|  Bob|
         +---+-----+
         """
-        if not isinstance(colsMap, dict):
-            raise PySparkTypeError(
-                error_class="NOT_DICT",
-                message_parameters={"arg_name": "colsMap", "arg_type": type(colsMap).__name__},
-            )
+        ...
 
-        col_names: List[str] = []
-        new_col_names: List[str] = []
-        for k, v in colsMap.items():
-            col_names.append(k)
-            new_col_names.append(v)
-
-        return DataFrame(
-            self._jdf.withColumnsRenamed(
-                _to_seq(self._sc, col_names), _to_seq(self._sc, new_col_names)
-            ),
-            self.sparkSession,
-        )
-
+    @dispatch_df_method
     def withMetadata(self, columnName: str, metadata: Dict[str, Any]) -> "DataFrame":
         """Returns a new :class:`DataFrame` by updating an existing column with metadata.
 
@@ -6402,18 +5532,7 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df_meta.schema['age'].metadata
         {'foo': 'bar'}
         """
-        from py4j.java_gateway import JVMView
-
-        if not isinstance(metadata, dict):
-            raise PySparkTypeError(
-                error_class="NOT_DICT",
-                message_parameters={"arg_name": "metadata", "arg_type": type(metadata).__name__},
-            )
-        sc = get_active_spark_context()
-        jmeta = cast(JVMView, sc._jvm).org.apache.spark.sql.types.Metadata.fromJson(
-            json.dumps(metadata)
-        )
-        return DataFrame(self._jdf.withMetadata(columnName, jmeta), self.sparkSession)
+        ...
 
     @overload
     def drop(self, cols: "ColumnOrName") -> "DataFrame":
@@ -6423,7 +5542,8 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def drop(self, *cols: str) -> "DataFrame":
         ...
 
-    def drop(self, *cols: "ColumnOrName") -> "DataFrame":  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def drop(self, *cols: "ColumnOrName") -> "DataFrame":
         """
         Returns a new :class:`DataFrame` without specified columns.
         This is a no-op if the schema doesn't contain the given column name(s).
@@ -6553,29 +5673,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         | 16|  Bob|    1|
         +---+-----+-----+
         """
-        column_names: List[str] = []
-        java_columns: List["JavaObject"] = []
+        ...
 
-        for c in cols:
-            if isinstance(c, str):
-                column_names.append(c)
-            elif isinstance(c, Column):
-                java_columns.append(c._jc)
-            else:
-                raise PySparkTypeError(
-                    error_class="NOT_COLUMN_OR_STR",
-                    message_parameters={"arg_name": "col", "arg_type": type(c).__name__},
-                )
-
-        jdf = self._jdf
-        if len(java_columns) > 0:
-            first_column, *remaining_columns = java_columns
-            jdf = jdf.drop(first_column, self._jseq(remaining_columns))
-        if len(column_names) > 0:
-            jdf = jdf.drop(self._jseq(column_names))
-
-        return DataFrame(jdf, self.sparkSession)
-
+    @dispatch_df_method
     def toDF(self, *cols: str) -> "DataFrame":
         """Returns a new :class:`DataFrame` that with new specified column names
 
@@ -6609,15 +5709,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         | 16|  Bob|
         +---+-----+
         """
-        for col in cols:
-            if not isinstance(col, str):
-                raise PySparkTypeError(
-                    error_class="NOT_LIST_OF_STR",
-                    message_parameters={"arg_name": "cols", "arg_type": type(col).__name__},
-                )
-        jdf = self._jdf.toDF(self._jseq(cols))
-        return DataFrame(jdf, self.sparkSession)
+        ...
 
+    @dispatch_df_method
     def transform(self, func: Callable[..., "DataFrame"], *args: Any, **kwargs: Any) -> "DataFrame":
         """Returns a new :class:`DataFrame`. Concise syntax for chaining custom transformations.
 
@@ -6673,12 +5767,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         | 13| 13.0|
         +---+-----+
         """
-        result = func(self, *args, **kwargs)
-        assert isinstance(
-            result, DataFrame
-        ), "Func returned an instance of type [%s], " "should have been DataFrame." % type(result)
-        return result
+        ...
 
+    @dispatch_df_method
     def sameSemantics(self, other: "DataFrame") -> bool:
         """
         Returns `True` when the logical query plans inside both :class:`DataFrame`\\s are equal and
@@ -6721,13 +5812,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> df1.withColumn("col1", df1.id * 2).sameSemantics(df2.withColumn("col0", df2.id * 2))
         True
         """
-        if not isinstance(other, DataFrame):
-            raise PySparkTypeError(
-                error_class="NOT_DATAFRAME",
-                message_parameters={"arg_name": "other", "arg_type": type(other).__name__},
-            )
-        return self._jdf.sameSemantics(other._jdf)
+        ...
 
+    @dispatch_df_method
     def semanticHash(self) -> int:
         """
         Returns a hash code of the logical query plan against this :class:`DataFrame`.
@@ -6756,8 +5843,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         >>> spark.range(10).selectExpr("id as col1").semanticHash()  # doctest: +SKIP
         1855039936
         """
-        return self._jdf.semanticHash()
+        ...
 
+    @dispatch_df_method
     def inputFiles(self) -> List[str]:
         """
         Returns a best-effort snapshot of the files that compose this :class:`DataFrame`.
@@ -6791,15 +5879,16 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...     len(df.inputFiles())
         1
         """
-        return list(self._jdf.inputFiles())
+        ...
 
+    @dispatch_df_method
     def where(self, condition: "ColumnOrName") -> "DataFrame":
         """
         :func:`where` is an alias for :func:`filter`.
 
         .. versionadded:: 1.3.0
         """
-        return self.filter(condition)
+        ...
 
     # Two aliases below were added for pandas compatibility many years ago.
     # There are too many differences compared to pandas and we cannot just
@@ -6814,22 +5903,25 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
     def groupby(self, __cols: Union[List[Column], List[str], List[int]]) -> "GroupedData":
         ...
 
-    def groupby(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def groupby(self, *cols: "ColumnOrNameOrOrdinal") -> "GroupedData":
         """
         :func:`groupby` is an alias for :func:`groupBy`.
 
         .. versionadded:: 1.4.0
         """
-        return self.groupBy(*cols)
+        ...
 
+    @dispatch_df_method
     def drop_duplicates(self, subset: Optional[List[str]] = None) -> "DataFrame":
         """
         :func:`drop_duplicates` is an alias for :func:`dropDuplicates`.
 
         .. versionadded:: 1.4.0
         """
-        return self.dropDuplicates(subset)
+        ...
 
+    @dispatch_df_method
     def writeTo(self, table: str) -> DataFrameWriterV2:
         """
         Create a write configuration builder for v2 sources.
@@ -6862,8 +5954,9 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         ...     "catalog.db.table"
         ... ).partitionedBy("col").createOrReplace()
         """
-        return DataFrameWriterV2(self, table)
+        ...
 
+    @dispatch_df_method
     def pandas_api(
         self, index_col: Optional[Union[str, List[str]]] = None
     ) -> "PandasOnSparkDataFrame":
@@ -6914,25 +6007,223 @@ class DataFrame(PandasMapOpsMixin, PandasConversionMixin):
         23   Alice
         16     Bob
         """
-        from pyspark.pandas.namespace import _get_index_map
-        from pyspark.pandas.frame import DataFrame as PandasOnSparkDataFrame
-        from pyspark.pandas.internal import InternalFrame
+        ...
 
-        index_spark_columns, index_names = _get_index_map(self, index_col)
-        internal = InternalFrame(
-            spark_frame=self,
-            index_spark_columns=index_spark_columns,
-            index_names=index_names,  # type: ignore[arg-type]
-        )
-        return PandasOnSparkDataFrame(internal)
+    @dispatch_df_method
+    def mapInPandas(
+        self,
+        func: "PandasMapIterFunction",
+        schema: Union[StructType, str],
+        barrier: bool = False,
+        profile: Optional[ResourceProfile] = None,
+    ) -> "DataFrame":
+        """
+        Maps an iterator of batches in the current :class:`DataFrame` using a Python native
+        function that is performed on pandas DataFrames both as input and output,
+        and returns the result as a :class:`DataFrame`.
+
+        This method applies the specified Python function to an iterator of
+        `pandas.DataFrame`\\s, each representing a batch of rows from the original DataFrame.
+        The returned iterator of `pandas.DataFrame`\\s are combined as a :class:`DataFrame`.
+        The size of the function's input and output can be different. Each `pandas.DataFrame`
+        size can be controlled by `spark.sql.execution.arrow.maxRecordsPerBatch`.
+
+        .. versionadded:: 3.0.0
+
+        .. versionchanged:: 3.4.0
+            Supports Spark Connect.
+
+        Parameters
+        ----------
+        func : function
+            a Python native function that takes an iterator of `pandas.DataFrame`\\s, and
+            outputs an iterator of `pandas.DataFrame`\\s.
+        schema : :class:`pyspark.sql.types.DataType` or str
+            the return type of the `func` in PySpark. The value can be either a
+            :class:`pyspark.sql.types.DataType` object or a DDL-formatted type string.
+        barrier : bool, optional, default False
+            Use barrier mode execution, ensuring that all Python workers in the stage will be
+            launched concurrently.
+
+            .. versionadded: 3.5.0
+
+        profile : :class:`pyspark.resource.ResourceProfile`. The optional ResourceProfile
+            to be used for mapInPandas.
+
+            .. versionadded: 4.0.0
 
 
-def _to_scala_map(sc: "SparkContext", jm: Dict) -> "JavaObject":
-    """
-    Convert a dict into a JVM Map.
-    """
-    assert sc._jvm is not None
-    return sc._jvm.PythonUtils.toScalaMap(jm)
+        Examples
+        --------
+        >>> df = spark.createDataFrame([(1, 21), (2, 30)], ("id", "age"))
+
+        Filter rows with id equal to 1:
+
+        >>> def filter_func(iterator):
+        ...     for pdf in iterator:
+        ...         yield pdf[pdf.id == 1]
+        ...
+        >>> df.mapInPandas(filter_func, df.schema).show()  # doctest: +SKIP
+        +---+---+
+        | id|age|
+        +---+---+
+        |  1| 21|
+        +---+---+
+
+        Compute the mean age for each id:
+
+        >>> def mean_age(iterator):
+        ...     for pdf in iterator:
+        ...         yield pdf.groupby("id").mean().reset_index()
+        ...
+        >>> df.mapInPandas(mean_age, "id: bigint, age: double").show()  # doctest: +SKIP
+        +---+----+
+        | id| age|
+        +---+----+
+        |  1|21.0|
+        |  2|30.0|
+        +---+----+
+
+        Add a new column with the double of the age:
+
+        >>> def double_age(iterator):
+        ...     for pdf in iterator:
+        ...         pdf["double_age"] = pdf["age"] * 2
+        ...         yield pdf
+        ...
+        >>> df.mapInPandas(
+        ...     double_age, "id: bigint, age: bigint, double_age: bigint").show()  # doctest: +SKIP
+        +---+---+----------+
+        | id|age|double_age|
+        +---+---+----------+
+        |  1| 21|        42|
+        |  2| 30|        60|
+        +---+---+----------+
+
+        Set ``barrier`` to ``True`` to force the ``mapInPandas`` stage running in the
+        barrier mode, it ensures all Python workers in the stage will be
+        launched concurrently.
+
+        >>> df.mapInPandas(filter_func, df.schema, barrier=True).show()  # doctest: +SKIP
+        +---+---+
+        | id|age|
+        +---+---+
+        |  1| 21|
+        +---+---+
+
+        Notes
+        -----
+        This API is experimental
+
+        See Also
+        --------
+        pyspark.sql.functions.pandas_udf
+        """
+        ...
+
+    @dispatch_df_method
+    def mapInArrow(
+        self,
+        func: "ArrowMapIterFunction",
+        schema: Union[StructType, str],
+        barrier: bool = False,
+        profile: Optional[ResourceProfile] = None,
+    ) -> "DataFrame":
+        """
+        Maps an iterator of batches in the current :class:`DataFrame` using a Python native
+        function that is performed on `pyarrow.RecordBatch`\\s both as input and output,
+        and returns the result as a :class:`DataFrame`.
+
+        This method applies the specified Python function to an iterator of
+        `pyarrow.RecordBatch`\\s, each representing a batch of rows from the original DataFrame.
+        The returned iterator of `pyarrow.RecordBatch`\\s are combined as a :class:`DataFrame`.
+        The size of the function's input and output can be different. Each `pyarrow.RecordBatch`
+        size can be controlled by `spark.sql.execution.arrow.maxRecordsPerBatch`.
+
+        .. versionadded:: 3.3.0
+
+        Parameters
+        ----------
+        func : function
+            a Python native function that takes an iterator of `pyarrow.RecordBatch`\\s, and
+            outputs an iterator of `pyarrow.RecordBatch`\\s.
+        schema : :class:`pyspark.sql.types.DataType` or str
+            the return type of the `func` in PySpark. The value can be either a
+            :class:`pyspark.sql.types.DataType` object or a DDL-formatted type string.
+        barrier : bool, optional, default False
+            Use barrier mode execution, ensuring that all Python workers in the stage will be
+            launched concurrently.
+
+            .. versionadded: 3.5.0
+
+        profile : :class:`pyspark.resource.ResourceProfile`. The optional ResourceProfile
+            to be used for mapInArrow.
+
+            .. versionadded: 4.0.0
+
+        Examples
+        --------
+        >>> import pyarrow  # doctest: +SKIP
+        >>> df = spark.createDataFrame([(1, 21), (2, 30)], ("id", "age"))
+        >>> def filter_func(iterator):
+        ...     for batch in iterator:
+        ...         pdf = batch.to_pandas()
+        ...         yield pyarrow.RecordBatch.from_pandas(pdf[pdf.id == 1])
+        >>> df.mapInArrow(filter_func, df.schema).show()  # doctest: +SKIP
+        +---+---+
+        | id|age|
+        +---+---+
+        |  1| 21|
+        +---+---+
+
+        Set ``barrier`` to ``True`` to force the ``mapInArrow`` stage running in the
+        barrier mode, it ensures all Python workers in the stage will be
+        launched concurrently.
+
+        >>> df.mapInArrow(filter_func, df.schema, barrier=True).show()  # doctest: +SKIP
+        +---+---+
+        | id|age|
+        +---+---+
+        |  1| 21|
+        +---+---+
+
+        Notes
+        -----
+        This API is unstable, and for developers.
+
+        See Also
+        --------
+        pyspark.sql.functions.pandas_udf
+        pyspark.sql.DataFrame.mapInPandas
+        """
+        ...
+
+    def toPandas(self) -> "PandasDataFrameLike":
+        """
+        Returns the contents of this :class:`DataFrame` as Pandas ``pandas.DataFrame``.
+
+        This is only available if Pandas is installed and available.
+
+        .. versionadded:: 1.3.0
+
+        .. versionchanged:: 3.4.0
+            Supports Spark Connect.
+
+        Notes
+        -----
+        This method should only be used if the resulting Pandas ``pandas.DataFrame`` is
+        expected to be small, as all the data is loaded into the driver's memory.
+
+        Usage with ``spark.sql.execution.arrow.pyspark.enabled=True`` is experimental.
+
+        Examples
+        --------
+        >>> df.toPandas()  # doctest: +SKIP
+           age   name
+        0    2  Alice
+        1    5    Bob
+        """
+        ...
 
 
 class DataFrameNaFunctions:
@@ -6947,13 +6238,14 @@ class DataFrameNaFunctions:
     def __init__(self, df: DataFrame):
         self.df = df
 
+    @dispatch_df_method
     def drop(
         self,
         how: str = "any",
         thresh: Optional[int] = None,
         subset: Optional[Union[str, Tuple[str, ...], List[str]]] = None,
     ) -> DataFrame:
-        return self.df.dropna(how=how, thresh=thresh, subset=subset)
+        ...
 
     drop.__doc__ = DataFrame.dropna.__doc__
 
@@ -6965,12 +6257,13 @@ class DataFrameNaFunctions:
     def fill(self, value: Dict[str, "LiteralType"]) -> DataFrame:
         ...
 
+    @dispatch_df_method
     def fill(
         self,
         value: Union["LiteralType", Dict[str, "LiteralType"]],
         subset: Optional[List[str]] = None,
     ) -> DataFrame:
-        return self.df.fillna(value=value, subset=subset)  # type: ignore[arg-type]
+        ...
 
     fill.__doc__ = DataFrame.fillna.__doc__
 
@@ -7000,7 +6293,8 @@ class DataFrameNaFunctions:
     ) -> DataFrame:
         ...
 
-    def replace(  # type: ignore[misc]
+    @dispatch_df_method  # type: ignore[misc]
+    def replace(
         self,
         to_replace: Union[List["LiteralType"], Dict["LiteralType", "OptionalPrimitiveType"]],
         value: Optional[
@@ -7008,7 +6302,7 @@ class DataFrameNaFunctions:
         ] = _NoValue,
         subset: Optional[List[str]] = None,
     ) -> DataFrame:
-        return self.df.replace(to_replace, value, subset)  # type: ignore[arg-type]
+        ...
 
     replace.__doc__ = DataFrame.replace.__doc__
 
@@ -7043,61 +6337,45 @@ class DataFrameStatFunctions:
     ) -> List[List[float]]:
         ...
 
+    @dispatch_df_method
     def approxQuantile(
         self,
         col: Union[str, List[str], Tuple[str]],
         probabilities: Union[List[float], Tuple[float]],
         relativeError: float,
     ) -> Union[List[float], List[List[float]]]:
-        return self.df.approxQuantile(col, probabilities, relativeError)
+        ...
 
     approxQuantile.__doc__ = DataFrame.approxQuantile.__doc__
 
+    @dispatch_df_method
     def corr(self, col1: str, col2: str, method: Optional[str] = None) -> float:
-        return self.df.corr(col1, col2, method)
+        ...
 
     corr.__doc__ = DataFrame.corr.__doc__
 
+    @dispatch_df_method
     def cov(self, col1: str, col2: str) -> float:
-        return self.df.cov(col1, col2)
+        ...
 
     cov.__doc__ = DataFrame.cov.__doc__
 
+    @dispatch_df_method
     def crosstab(self, col1: str, col2: str) -> DataFrame:
-        return self.df.crosstab(col1, col2)
+        ...
 
     crosstab.__doc__ = DataFrame.crosstab.__doc__
 
+    @dispatch_df_method
     def freqItems(self, cols: List[str], support: Optional[float] = None) -> DataFrame:
-        return self.df.freqItems(cols, support)
+        ...
 
     freqItems.__doc__ = DataFrame.freqItems.__doc__
 
+    @dispatch_df_method
     def sampleBy(
         self, col: str, fractions: Dict[Any, float], seed: Optional[int] = None
     ) -> DataFrame:
-        return self.df.sampleBy(col, fractions, seed)
+        ...
 
     sampleBy.__doc__ = DataFrame.sampleBy.__doc__
-
-
-def _test() -> None:
-    import doctest
-    from pyspark.sql import SparkSession
-    import pyspark.sql.dataframe
-
-    globs = pyspark.sql.dataframe.__dict__.copy()
-    spark = SparkSession.builder.master("local[4]").appName("sql.dataframe tests").getOrCreate()
-    globs["spark"] = spark
-    (failure_count, test_count) = doctest.testmod(
-        pyspark.sql.dataframe,
-        globs=globs,
-        optionflags=doctest.ELLIPSIS | doctest.NORMALIZE_WHITESPACE | doctest.REPORT_NDIFF,
-    )
-    spark.stop()
-    if failure_count:
-        sys.exit(-1)
-
-
-if __name__ == "__main__":
-    _test()
