+++ python/pyspark/sql/pandas/map_ops.py
@@ -41,108 +41,6 @@ class PandasMapOpsMixin:
         barrier: bool = False,
         profile: Optional[ResourceProfile] = None,
     ) -> "DataFrame":
-        """
-        Maps an iterator of batches in the current :class:`DataFrame` using a Python native
-        function that is performed on pandas DataFrames both as input and output,
-        and returns the result as a :class:`DataFrame`.
-
-        This method applies the specified Python function to an iterator of
-        `pandas.DataFrame`\\s, each representing a batch of rows from the original DataFrame.
-        The returned iterator of `pandas.DataFrame`\\s are combined as a :class:`DataFrame`.
-        The size of the function's input and output can be different. Each `pandas.DataFrame`
-        size can be controlled by `spark.sql.execution.arrow.maxRecordsPerBatch`.
-
-        .. versionadded:: 3.0.0
-
-        .. versionchanged:: 3.4.0
-            Supports Spark Connect.
-
-        Parameters
-        ----------
-        func : function
-            a Python native function that takes an iterator of `pandas.DataFrame`\\s, and
-            outputs an iterator of `pandas.DataFrame`\\s.
-        schema : :class:`pyspark.sql.types.DataType` or str
-            the return type of the `func` in PySpark. The value can be either a
-            :class:`pyspark.sql.types.DataType` object or a DDL-formatted type string.
-        barrier : bool, optional, default False
-            Use barrier mode execution, ensuring that all Python workers in the stage will be
-            launched concurrently.
-
-            .. versionadded: 3.5.0
-
-        profile : :class:`pyspark.resource.ResourceProfile`. The optional ResourceProfile
-            to be used for mapInPandas.
-
-            .. versionadded: 4.0.0
-
-
-        Examples
-        --------
-        >>> df = spark.createDataFrame([(1, 21), (2, 30)], ("id", "age"))
-
-        Filter rows with id equal to 1:
-
-        >>> def filter_func(iterator):
-        ...     for pdf in iterator:
-        ...         yield pdf[pdf.id == 1]
-        ...
-        >>> df.mapInPandas(filter_func, df.schema).show()  # doctest: +SKIP
-        +---+---+
-        | id|age|
-        +---+---+
-        |  1| 21|
-        +---+---+
-
-        Compute the mean age for each id:
-
-        >>> def mean_age(iterator):
-        ...     for pdf in iterator:
-        ...         yield pdf.groupby("id").mean().reset_index()
-        ...
-        >>> df.mapInPandas(mean_age, "id: bigint, age: double").show()  # doctest: +SKIP
-        +---+----+
-        | id| age|
-        +---+----+
-        |  1|21.0|
-        |  2|30.0|
-        +---+----+
-
-        Add a new column with the double of the age:
-
-        >>> def double_age(iterator):
-        ...     for pdf in iterator:
-        ...         pdf["double_age"] = pdf["age"] * 2
-        ...         yield pdf
-        ...
-        >>> df.mapInPandas(
-        ...     double_age, "id: bigint, age: bigint, double_age: bigint").show()  # doctest: +SKIP
-        +---+---+----------+
-        | id|age|double_age|
-        +---+---+----------+
-        |  1| 21|        42|
-        |  2| 30|        60|
-        +---+---+----------+
-
-        Set ``barrier`` to ``True`` to force the ``mapInPandas`` stage running in the
-        barrier mode, it ensures all Python workers in the stage will be
-        launched concurrently.
-
-        >>> df.mapInPandas(filter_func, df.schema, barrier=True).show()  # doctest: +SKIP
-        +---+---+
-        | id|age|
-        +---+---+
-        |  1| 21|
-        +---+---+
-
-        Notes
-        -----
-        This API is experimental
-
-        See Also
-        --------
-        pyspark.sql.functions.pandas_udf
-        """
         from pyspark.sql import DataFrame
         from pyspark.sql.pandas.functions import pandas_udf
 
@@ -165,73 +63,6 @@ class PandasMapOpsMixin:
         barrier: bool = False,
         profile: Optional[ResourceProfile] = None,
     ) -> "DataFrame":
-        """
-        Maps an iterator of batches in the current :class:`DataFrame` using a Python native
-        function that is performed on `pyarrow.RecordBatch`\\s both as input and output,
-        and returns the result as a :class:`DataFrame`.
-
-        This method applies the specified Python function to an iterator of
-        `pyarrow.RecordBatch`\\s, each representing a batch of rows from the original DataFrame.
-        The returned iterator of `pyarrow.RecordBatch`\\s are combined as a :class:`DataFrame`.
-        The size of the function's input and output can be different. Each `pyarrow.RecordBatch`
-        size can be controlled by `spark.sql.execution.arrow.maxRecordsPerBatch`.
-
-        .. versionadded:: 3.3.0
-
-        Parameters
-        ----------
-        func : function
-            a Python native function that takes an iterator of `pyarrow.RecordBatch`\\s, and
-            outputs an iterator of `pyarrow.RecordBatch`\\s.
-        schema : :class:`pyspark.sql.types.DataType` or str
-            the return type of the `func` in PySpark. The value can be either a
-            :class:`pyspark.sql.types.DataType` object or a DDL-formatted type string.
-        barrier : bool, optional, default False
-            Use barrier mode execution, ensuring that all Python workers in the stage will be
-            launched concurrently.
-
-            .. versionadded: 3.5.0
-
-        profile : :class:`pyspark.resource.ResourceProfile`. The optional ResourceProfile
-            to be used for mapInArrow.
-
-            .. versionadded: 4.0.0
-
-        Examples
-        --------
-        >>> import pyarrow  # doctest: +SKIP
-        >>> df = spark.createDataFrame([(1, 21), (2, 30)], ("id", "age"))
-        >>> def filter_func(iterator):
-        ...     for batch in iterator:
-        ...         pdf = batch.to_pandas()
-        ...         yield pyarrow.RecordBatch.from_pandas(pdf[pdf.id == 1])
-        >>> df.mapInArrow(filter_func, df.schema).show()  # doctest: +SKIP
-        +---+---+
-        | id|age|
-        +---+---+
-        |  1| 21|
-        +---+---+
-
-        Set ``barrier`` to ``True`` to force the ``mapInArrow`` stage running in the
-        barrier mode, it ensures all Python workers in the stage will be
-        launched concurrently.
-
-        >>> df.mapInArrow(filter_func, df.schema, barrier=True).show()  # doctest: +SKIP
-        +---+---+
-        | id|age|
-        +---+---+
-        |  1| 21|
-        +---+---+
-
-        Notes
-        -----
-        This API is unstable, and for developers.
-
-        See Also
-        --------
-        pyspark.sql.functions.pandas_udf
-        pyspark.sql.DataFrame.mapInPandas
-        """
         from pyspark.sql import DataFrame
         from pyspark.sql.pandas.functions import pandas_udf
